match test directories to src, index quota test
