+++ test/cli/commands/smoke-test/smoke-tests/aws/resource-tests/eip-smoke-tests.test.ts
