+++ test/cli/commands/smoke-test/smoke-tests/aws/resource-tests/s3-smoke-tests.test.ts
