+++ test/cli/commands/smoke-test/smoke-tests/aws/resource-tests/sqs-smoke-tests.test.ts
